use safer comparison function

+++ src/padding.c
@@ -9,7 +9,7 @@ unsigned add_padding(Args* args) {
   }
 
   for (int i = 0; i < args->padding; ++i) {
-    unsigned cat_len = length + (i + 2) * args->p_string_len;
+    size_t cat_len = length + (i + 2) * args->p_string_len;
     strlcat(args->string, args->padding_string, cat_len);
   }
   return length + (args->padding * args->p_string_len);
